requirements.txt
# ========== FastAPI Backend ==========
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# ========== Machine Learning (Core) ==========
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# ========== Deep Learning (Optional) ==========
# Uncomment to enable Deep Learning support
# tensorflow==2.15.0
# keras==2.15.0

# ========== Feature Extraction ==========
tldextract==5.1.1

# ========== Database ==========
sqlalchemy==2.0.23

# ========== Authentication ==========
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# ========== External Datasets ==========
datasets==3.1.0
huggingface-hub==0.26.2

# ========== Visualization (Optional) ==========
# matplotlib==3.7.2
# seaborn==0.12.2
